Added class based YOLOv5
